requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
xgboost
Pillow
plotly
joblib 